Time: 65 ms (71.20%), Space: 5.9 MB (47.83%) - LeetHub

+++ 1680-concatenation-of-consecutive-binary-numbers/1680-concatenation-of-consecutive-binary-numbers.cpp
@@ -1,5 +1,5 @@
 using ll = long long;
-const int mod = 1e9+7;
+const ll mod = 1e9+7;
 class Solution {
 public:
     int concatenatedBinary(int n) {
@@ -11,4 +11,4 @@ public:
         }
         return ans;
     }
-}; 
+};
